# [POAE7-2921] Enable Vectorized Filter and Project for TPC-H Q6