--- a/cpp/src/cider/exec/nextgen/jitlib/llvmjit/LLVMJITValue.h
+++ b/cpp/src/cider/exec/nextgen/jitlib/llvmjit/LLVMJITValue.h
@@ -53,6 +53,7 @@ class LLVMJITValue final : public JITValue {
 
   JITValuePointer andOp(JITValue& rh) override;
   JITValuePointer orOp(JITValue& rh) override;
+  JITValuePointer xorOp(JITValue& rh) override;
   JITValuePointer notOp() override;
 
   JITValuePointer add(JITValue& rh) override;