--- a/cpp/src/cider/exec/nextgen/utils/TypeUtils.h
+++ b/cpp/src/cider/exec/nextgen/utils/TypeUtils.h
@@ -157,6 +157,7 @@ inline bool isVectorizableType(SQLTypes type) {
     case kFLOAT:
     case kDOUBLE:
     case kBOOLEAN:
+    case kDATE:
       return true;
     default:
       return false;