--- a/cpp/src/cider/type/plan/Expr.h
+++ b/cpp/src/cider/type/plan/Expr.h
@@ -210,6 +210,8 @@ class Expr : public std::enable_shared_from_this<Expr> {
 
   void setTrivialNullProcess(bool flag) { is_trivial_null_process_ = flag; }
 
+  void setGenerateVectorizedBoolCode(bool flag) { generate_vectorized_bool_code_ = flag; }
+
  protected:
   JITTypeTag getJITTag(const SQLTypes& st) {
     return cider::exec::nextgen::utils::getJITTypeTag(st);
@@ -224,6 +226,7 @@ class Expr : public std::enable_shared_from_this<Expr> {
   bool auto_vectorizable_{false};
   bool is_column_var_{false};
   bool is_trivial_null_process_{true};
+  bool generate_vectorized_bool_code_{false};
 };
 
 using ExpressionPtr = std::shared_ptr<Analyzer::Expr>;