--- a/cpp/src/cider/type/plan/Expr.h
+++ b/cpp/src/cider/type/plan/Expr.h
@@ -206,6 +206,10 @@ class Expr : public std::enable_shared_from_this<Expr> {
 
   void setNullable(bool nullable) { type_info.set_notnull(!nullable); }
 
+  bool isTrivialNullProcess() const { return is_trivial_null_process_; }
+
+  void setTrivialNullProcess(bool flag) { is_trivial_null_process_ = flag; }
+
  protected:
   JITTypeTag getJITTag(const SQLTypes& st) {
     return cider::exec::nextgen::utils::getJITTypeTag(st);
@@ -219,6 +223,7 @@ class Expr : public std::enable_shared_from_this<Expr> {
   size_t local_index_{0};  // 1-based index of input column in CodegenContext.
   bool auto_vectorizable_{false};
   bool is_column_var_{false};
+  bool is_trivial_null_process_{true};
 };
 
 using ExpressionPtr = std::shared_ptr<Analyzer::Expr>;